[#IC-298] extract getMessages fallback implementation
