+++ GetMessages/handler.ts
@@ -1,9 +1,3 @@
-import {
-  asyncIteratorToPageArray,
-  flattenAsyncIterator,
-  mapAsyncIterator
-} from "@pagopa/io-functions-commons/dist/src/utils/async";
-import { retrievedMessageToPublic } from "@pagopa/io-functions-commons/dist/src/utils/messages";
 import { FiscalCodeMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/fiscalcode";
 import {
   withRequestMiddlewares,
@@ -13,23 +7,15 @@ import {
   IResponseErrorQuery,
   ResponseErrorQuery
 } from "@pagopa/io-functions-commons/dist/src/utils/response";
-import {
-  PageResults,
-  toPageResults
-} from "@pagopa/io-functions-commons/dist/src/utils/paging";
+import { PageResults } from "@pagopa/io-functions-commons/dist/src/utils/paging";
 import {
   defaultPageSize,
-  MessageModel,
-  RetrievedMessage
+  MessageModel
 } from "@pagopa/io-functions-commons/dist/src/models/message";
-import { CreatedMessageWithoutContent } from "@pagopa/io-functions-commons/dist/generated/definitions/CreatedMessageWithoutContent";
 import { OptionalQueryParamMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/optional_query_param";
 import * as express from "express";
-import * as A from "fp-ts/lib/Array";
-import * as E from "fp-ts/lib/Either";
-import { flow, pipe } from "fp-ts/lib/function";
+import { pipe } from "fp-ts/lib/function";
 import * as TE from "fp-ts/lib/TaskEither";
-import * as t from "io-ts";
 import { ServiceModel } from "@pagopa/io-functions-commons/dist/src/models/service";
 import { BooleanFromString } from "@pagopa/ts-commons/lib/booleans";
 import {
@@ -49,18 +35,8 @@ import * as O from "fp-ts/lib/Option";
 import { ContextMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/context_middleware";
 import { Context } from "@azure/functions";
 import { RedisClient } from "redis";
-import {
-  CreatedMessageWithoutContentWithStatus,
-  enrichMessagesData,
-  enrichMessagesStatus
-} from "../utils/messages";
 import { MessageStatusExtendedQueryModel } from "../model/message_status_query";
-
-type RetrievedNotPendingMessage = t.TypeOf<typeof RetrievedNotPendingMessage>;
-const RetrievedNotPendingMessage = t.intersection([
-  RetrievedMessage,
-  t.interface({ isPending: t.literal(false) })
-]);
+import { getMessagesFromFallback } from "./getMessages.fallback";
 
 type IGetMessagesHandlerResponse =
   | IResponseSuccessJson<PageResults>
@@ -84,23 +60,6 @@ type IGetMessagesHandler = (
   maybeMaximumId: O.Option<NonEmptyString>,
   maybeMinimumId: O.Option<NonEmptyString>
 ) => Promise<IGetMessagesHandlerResponse>;
-
-const filterMessages = (shouldGetArchivedMessages: boolean) => (
-  // eslint-disable-next-line functional/prefer-readonly-type, @typescript-eslint/array-type
-  messages: E.Either<Error, CreatedMessageWithoutContentWithStatus>[]
-  // eslint-disable-next-line functional/prefer-readonly-type, @typescript-eslint/array-type
-): E.Either<Error, CreatedMessageWithoutContentWithStatus>[] =>
-  pipe(
-    messages,
-    A.filter(
-      flow(
-        // never filter away errors
-        E.mapLeft(() => true),
-        E.map(mess => mess.is_archived === shouldGetArchivedMessages),
-        E.toUnion
-      )
-    )
-  );
 
 /**
  * Handles requests for getting all message for a recipient.
@@ -144,79 +103,30 @@ export const GetMessagesHandler = (
         maximumId,
         minimumId
       }) =>
-        pipe(
-          messageModel.findMessages(fiscalCode, pageSize, maximumId, minimumId),
-          TE.mapLeft(failure => ResponseErrorQuery(failure.kind, failure)),
-          TE.map(i => mapAsyncIterator(i, A.rights)),
-          TE.map(i =>
-            mapAsyncIterator(i, A.filter(RetrievedNotPendingMessage.is))
-          ),
-          TE.map(i => mapAsyncIterator(i, A.map(retrievedMessageToPublic))),
-          TE.chainW(i =>
-            // check whether we should enrich messages or not
-            pipe(
-              TE.fromPredicate(
-                () => shouldEnrichResultData === true,
-                () =>
-                  // if no enrichment is requested we just wrap messages
-                  mapAsyncIterator(
-                    i,
-                    A.map(async e =>
-                      E.right<Error, CreatedMessageWithoutContent>(e)
-                    )
-                  )
-              )(i),
-              TE.map(j =>
-                mapAsyncIterator(j, async m =>
-                  enrichMessagesStatus(context, messageStatusModel)(m)()
-                )
-              ),
-              TE.map(j =>
-                mapAsyncIterator(j, filterMessages(shouldGetArchivedMessages))
-              ),
-              TE.map(j =>
-                mapAsyncIterator(j, x => [
-                  // Do not enrich messages if errors occurred
-                  ...pipe(
-                    A.lefts(x),
-                    A.map(async y => E.left(y))
-                  ),
-                  ...pipe(
-                    A.rights(x),
-                    enrichMessagesData(
-                      context,
-                      messageModel,
-                      serviceModel,
-                      blobService,
-                      redisClient,
-                      serviceCacheTtl
-                    )
-                  )
-                ])
-              ),
-              // we need to make a TaskEither of the Either[] mapped above
-              TE.orElse(TE.of),
-              TE.map(flattenAsyncIterator),
-              TE.chain(_ =>
-                TE.tryCatch(
-                  () => asyncIteratorToPageArray(_, pageSize),
-                  E.toError
-                )
-              ),
-              TE.chain(
-                TE.fromPredicate(
-                  page => !page.results.some(E.isLeft),
-                  () => new Error("Cannot enrich data")
-                )
-              ),
-              TE.map(({ hasMoreResults, results }) =>
-                toPageResults(A.rights([...results]), hasMoreResults)
-              ),
-              TE.mapLeft(e => ResponseErrorInternal(e.message))
-            )
-          )
-        )
+        getMessagesFromFallback(
+          messageModel,
+          messageStatusModel,
+          serviceModel,
+          blobService,
+          redisClient,
+          serviceCacheTtl
+        )({
+          context,
+          fiscalCode,
+          maximumId,
+          minimumId,
+          pageSize,
+          shouldEnrichResultData,
+          shouldGetArchivedMessages
+        })
     ),
+    TE.mapLeft(e => {
+      if (e instanceof Error) {
+        return ResponseErrorInternal(e.message);
+      } else {
+        return ResponseErrorQuery(e.kind, e);
+      }
+    }),
     TE.map(ResponseSuccessJson),
     TE.toUnion
   )();
